# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport Api Swagger Documentation Anatation",
        "description": "Sport Api Documentation to build the frontend of this site",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "abduvaliyevasad23@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Sport api server link"
        }
    ],
    "paths": {
...
        "/api/check": {
            "get": {
                "tags": [
                    "Check"
                ],
                "summary": "Get all data from fillials database",
                "description": "Via this link All fillials` datas come",
                "operationId": "check",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Check"
                ],
                "summary": "Register a pasport",
                "operationId": "storeCheck",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "direction_id": {
                                        "description": "direction ID",
                                        "type": "integer"
                                    },
                                    "direction_category_id": {
                                        "description": "Direction category ID",
                                        "type": "integer"
                                    },
                                    "score": {
                                        "description": "Direction category score",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "direction_category_name": {
                                        "description": "Direction category name",
                                        "type": "string"
                                    },
                                    "direction_category_name_answer": {
                                        "description": "Direction category answer",
                                        "type": "string"
                                    },
                                    "pdf": {
                                        "description": "upload any pdf file",
                                        "type": "file"
                                    },
                                    "admin_permission": {
                                        "description": "Permission of one of an Admin",
                                        "type": "string"
                                    },
                                    "messages": {
                                        "description": "message of one of an Admin",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/check/{check}": {
            "get": {
                "tags": [
                    "Check"
                ],
                "summary": "Get one data from Avatar database",
                "description": "Via this link a Avatar`s data comes to show",
                "operationId": "showCheck",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "check",
                        "in": "path",
                        "description": "ID for avatar data",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/check/{check}/update": {
            "post": {
                "tags": [
                    "Check"
                ],
                "summary": "Updated",
                "description": "Update this avatar",
                "operationId": "updateCheck",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "check",
                        "in": "path",
                        "description": "avatar that to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "direction_category_name": {
                                        "description": "Direction category name",
                                        "type": "string"
                                    },
                                    "score": {
                                        "description": "score",
                                        "type": "integer"
                                    },
                                    "pdf": {
                                        "description": "upload any pdf file",
                                        "type": "file"
                                    },
                                    "admin_permission": {
                                        "description": "Permission of one of an Admin",
                                        "type": "string"
                                    },
                                    "messages": {
                                        "description": "message of one of an Admin",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid avatar supplied"
                    },
                    "404": {
                        "description": "avatar not found"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/check/{check}/delete": {
            "delete": {
                "tags": [
                    "Check"
                ],
                "summary": "Deletes a avatar",
                "operationId": "destroyCheck",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "check",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid course supplied"
                    },
                    "404": {
                        "description": "Avatar not found"
                    }
                }
            }
        },
...
        "/api/pasport": {
            "get": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Get all data from Pasport database",
                "description": "Via this link All Pasport` datas come",
                "operationId": "pasport",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Register a pasport",
                "operationId": "storePasport",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "pnfl": {
                                        "description": "Type your pnfl",
                                        "type": "integer"
                                    },
                                    "pasport_seria": {
                                        "description": "Type your pasport seria",
                                        "type": "string"
                                    },
                                    "pasport_seria_code": {
                                        "description": "Type your pasport seria code",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}